+++ src/components/Table/TableTh.vue
@@ -24,7 +24,6 @@ const props = withDefaults(defineProps<ITableTd>(), {
   font-weight: 700;
   font-size: var(--th-font-size, 14px);
   text-align: center;
-  line-height: 28px;
 
   border: none;
   border-left: 0.5px solid var(--border-color);